requirements.txt
flask
httpx
python-telegram-bot==21.6
